extract shared storage key codec
