+++ crates/jazz-tools/src/storage/opfs_btree.rs
@@ -30,7 +30,13 @@ use crate::object::{BranchName, ObjectId};
 use crate::query_manager::types::Value;
 use crate::sync_manager::PersistenceTier;
 
-use super::{LoadedBranch, Storage, StorageError, encode_value};
+use super::{
+    LoadedBranch, Storage, StorageError, encode_value,
+    key_codec::{
+        ack_key, branch_tips_key, commit_key, commit_prefix, increment_bytes, increment_string,
+        index_entry_key, index_prefix, index_value_prefix, obj_meta_key, parse_uuid_from_index_key,
+    },
+};
 
 const MIN_CACHE_SIZE_BYTES: usize = 4 * 1024 * 1024;
 
@@ -165,62 +171,6 @@ impl OpfsBTreeStorage {
         tracing::info!(obj_count, idx_count, ack_count, "OpfsBTreeStorage opened");
     }
 
-    fn obj_meta_key(id: ObjectId) -> String {
-        format!("obj:{}:meta", format_uuid(id))
-    }
-
-    fn branch_tips_key(object_id: ObjectId, branch: &BranchName) -> String {
-        format!("obj:{}:br:{}:tips", format_uuid(object_id), branch)
-    }
-
-    fn commit_key(object_id: ObjectId, branch: &BranchName, commit_id: CommitId) -> String {
-        format!(
-            "obj:{}:br:{}:c:{}",
-            format_uuid(object_id),
-            branch,
-            hex::encode(commit_id.0)
-        )
-    }
-
-    fn commit_prefix(object_id: ObjectId, branch: &BranchName) -> String {
-        format!("obj:{}:br:{}:c:", format_uuid(object_id), branch)
-    }
-
-    fn ack_key(commit_id: CommitId) -> String {
-        format!("ack:{}", hex::encode(commit_id.0))
-    }
-
-    fn index_entry_key(
-        table: &str,
-        column: &str,
-        branch: &str,
-        value: &Value,
-        row_id: ObjectId,
-    ) -> String {
-        format!(
-            "idx:{}:{}:{}:{}:{}",
-            table,
-            column,
-            branch,
-            hex::encode(encode_value(value)),
-            format_uuid(row_id)
-        )
-    }
-
-    fn index_value_prefix(table: &str, column: &str, branch: &str, value: &Value) -> String {
-        format!(
-            "idx:{}:{}:{}:{}:",
-            table,
-            column,
-            branch,
-            hex::encode(encode_value(value))
-        )
-    }
-
-    fn index_prefix(table: &str, column: &str, branch: &str) -> String {
-        format!("idx:{}:{}:{}:", table, column, branch)
-    }
-
     fn tree_insert(&self, key: &str, value: &[u8]) -> Result<(), StorageError> {
         self.with_tree_mut(|tree| tree.put(key.as_bytes(), value).map_err(map_storage_err))
     }
@@ -288,7 +238,7 @@ impl Storage for OpfsBTreeStorage {
         id: ObjectId,
         metadata: HashMap<String, String>,
     ) -> Result<(), StorageError> {
-        let key = Self::obj_meta_key(id);
+        let key = obj_meta_key(id);
         let json = serde_json::to_vec(&metadata)
             .map_err(|e| StorageError::IoError(format!("serialize metadata: {}", e)))?;
         self.tree_insert(&key, &json)
@@ -298,7 +248,7 @@ impl Storage for OpfsBTreeStorage {
         &self,
         id: ObjectId,
     ) -> Result<Option<HashMap<String, String>>, StorageError> {
-        let key = Self::obj_meta_key(id);
+        let key = obj_meta_key(id);
         match self.tree_read(&key)? {
             Some(data) => {
                 let meta: HashMap<String, String> = serde_json::from_slice(&data)
@@ -314,16 +264,16 @@ impl Storage for OpfsBTreeStorage {
         object_id: ObjectId,
         branch: &BranchName,
     ) -> Result<Option<LoadedBranch>, StorageError> {
-        let meta_key = Self::obj_meta_key(object_id);
+        let meta_key = obj_meta_key(object_id);
         if self.tree_read(&meta_key)?.is_none() {
             return Ok(None);
         }
 
-        let commit_prefix = Self::commit_prefix(object_id, branch);
+        let commit_prefix = commit_prefix(object_id, branch);
         let commit_entries = self.tree_scan_prefix(&commit_prefix)?;
 
         if commit_entries.is_empty() {
-            let tips_key = Self::branch_tips_key(object_id, branch);
+            let tips_key = branch_tips_key(object_id, branch);
             if self.tree_read(&tips_key)?.is_none() {
                 return Ok(None);
             }
@@ -334,7 +284,7 @@ impl Storage for OpfsBTreeStorage {
             let mut commit: Commit = serde_json::from_slice(data)
                 .map_err(|e| StorageError::IoError(format!("deserialize commit: {}", e)))?;
 
-            let ack_key = Self::ack_key(commit.id());
+            let ack_key = ack_key(commit.id());
             if let Some(ack_data) = self.tree_read(&ack_key)? {
                 let tiers: HashSet<PersistenceTier> = serde_json::from_slice(&ack_data)
                     .map_err(|e| StorageError::IoError(format!("deserialize ack: {}", e)))?;
@@ -344,7 +294,7 @@ impl Storage for OpfsBTreeStorage {
             commits.push(commit);
         }
 
-        let tips_key = Self::branch_tips_key(object_id, branch);
+        let tips_key = branch_tips_key(object_id, branch);
         let tails = match self.tree_read(&tips_key)? {
             Some(data) => serde_json::from_slice(&data)
                 .map_err(|e| StorageError::IoError(format!("deserialize tips: {}", e)))?,
@@ -362,12 +312,12 @@ impl Storage for OpfsBTreeStorage {
     ) -> Result<(), StorageError> {
         let commit_id = commit.id();
 
-        let commit_key = Self::commit_key(object_id, branch, commit_id);
+        let commit_key = commit_key(object_id, branch, commit_id);
         let commit_json = serde_json::to_vec(&commit)
             .map_err(|e| StorageError::IoError(format!("serialize commit: {}", e)))?;
         self.tree_insert(&commit_key, &commit_json)?;
 
-        let tips_key = Self::branch_tips_key(object_id, branch);
+        let tips_key = branch_tips_key(object_id, branch);
         let mut tips: HashSet<CommitId> = match self.tree_read(&tips_key)? {
             Some(data) => serde_json::from_slice(&data)
                 .map_err(|e| StorageError::IoError(format!("deserialize tips: {}", e)))?,
@@ -392,10 +342,10 @@ impl Storage for OpfsBTreeStorage {
         branch: &BranchName,
         commit_id: CommitId,
     ) -> Result<(), StorageError> {
-        let commit_key = Self::commit_key(object_id, branch, commit_id);
+        let commit_key = commit_key(object_id, branch, commit_id);
         self.tree_delete(&commit_key)?;
 
-        let tips_key = Self::branch_tips_key(object_id, branch);
+        let tips_key = branch_tips_key(object_id, branch);
         if let Some(data) = self.tree_read(&tips_key)? {
             let mut tips: HashSet<CommitId> = serde_json::from_slice(&data)
                 .map_err(|e| StorageError::IoError(format!("deserialize tips: {}", e)))?;
@@ -414,7 +364,7 @@ impl Storage for OpfsBTreeStorage {
         branch: &BranchName,
         tails: Option<HashSet<CommitId>>,
     ) -> Result<(), StorageError> {
-        let tips_key = Self::branch_tips_key(object_id, branch);
+        let tips_key = branch_tips_key(object_id, branch);
         match tails {
             Some(t) => {
                 let json = serde_json::to_vec(&t)
@@ -433,7 +383,7 @@ impl Storage for OpfsBTreeStorage {
         commit_id: CommitId,
         tier: PersistenceTier,
     ) -> Result<(), StorageError> {
-        let key = Self::ack_key(commit_id);
+        let key = ack_key(commit_id);
         let mut tiers: HashSet<PersistenceTier> = match self.tree_read(&key)? {
             Some(data) => serde_json::from_slice(&data)
                 .map_err(|e| StorageError::IoError(format!("deserialize ack: {}", e)))?,
@@ -454,7 +404,7 @@ impl Storage for OpfsBTreeStorage {
         row_id: ObjectId,
     ) -> Result<(), StorageError> {
         tracing::trace!(table, column, branch, ?row_id, "index_insert");
-        let key = Self::index_entry_key(table, column, branch, value, row_id);
+        let key = index_entry_key(table, column, branch, value, row_id);
         self.tree_insert(&key, &[0x01])
     }
 
@@ -467,7 +417,7 @@ impl Storage for OpfsBTreeStorage {
         row_id: ObjectId,
     ) -> Result<(), StorageError> {
         tracing::trace!(table, column, branch, ?row_id, "index_remove");
-        let key = Self::index_entry_key(table, column, branch, value, row_id);
+        let key = index_entry_key(table, column, branch, value, row_id);
         self.tree_delete(&key)
     }
 
@@ -479,7 +429,7 @@ impl Storage for OpfsBTreeStorage {
         value: &Value,
     ) -> Vec<ObjectId> {
         tracing::trace!(table, column, branch, "index_lookup");
-        let prefix = Self::index_value_prefix(table, column, branch, value);
+        let prefix = index_value_prefix(table, column, branch, value);
         match self.tree_scan_keys(&prefix) {
             Ok(keys) => keys
                 .iter()
@@ -497,7 +447,7 @@ impl Storage for OpfsBTreeStorage {
         start: Bound<&Value>,
         end: Bound<&Value>,
     ) -> Vec<ObjectId> {
-        let base_prefix = Self::index_prefix(table, column, branch);
+        let base_prefix = index_prefix(table, column, branch);
 
         let start_key = match start {
             Bound::Included(v) => {
@@ -543,7 +493,7 @@ impl Storage for OpfsBTreeStorage {
     }
 
     fn index_scan_all(&self, table: &str, column: &str, branch: &str) -> Vec<ObjectId> {
-        let prefix = Self::index_prefix(table, column, branch);
+        let prefix = index_prefix(table, column, branch);
         match self.tree_scan_keys(&prefix) {
             Ok(keys) => keys
                 .iter()
@@ -569,37 +519,6 @@ impl Storage for OpfsBTreeStorage {
 
 fn map_storage_err(error: BTreeError) -> StorageError {
     StorageError::IoError(format!("opfs-btree: {}", error))
-}
-
-fn format_uuid(id: ObjectId) -> String {
-    hex::encode(id.uuid().as_bytes())
-}
-
-fn parse_uuid_from_index_key(key: &str) -> Option<ObjectId> {
-    let uuid_hex = key.rsplit(':').next()?;
-    let bytes = hex::decode(uuid_hex).ok()?;
-    if bytes.len() != 16 {
-        return None;
-    }
-    let uuid = uuid::Uuid::from_bytes(bytes.try_into().ok()?);
-    Some(ObjectId(internment::Intern::new(uuid)))
-}
-
-fn increment_bytes(bytes: &mut Vec<u8>) {
-    for i in (0..bytes.len()).rev() {
-        if bytes[i] < 0xFF {
-            bytes[i] += 1;
-            bytes.truncate(i + 1);
-            return;
-        }
-    }
-    bytes.push(0x00);
-}
-
-fn increment_string(s: &mut String) {
-    let mut bytes = std::mem::take(s).into_bytes();
-    increment_bytes(&mut bytes);
-    *s = String::from_utf8(bytes).unwrap_or_default();
 }
 
 #[cfg(test)]
@@ -789,8 +708,8 @@ mod tests {
             .store_ack_tier(commit_id, PersistenceTier::EdgeServer)
             .unwrap();
 
-        let ack_key = OpfsBTreeStorage::ack_key(commit_id);
-        let data = storage.tree_read(&ack_key).unwrap().unwrap();
+        let key = ack_key(commit_id);
+        let data = storage.tree_read(&key).unwrap().unwrap();
         let tiers: HashSet<PersistenceTier> = serde_json::from_slice(&data).unwrap();
         assert!(tiers.contains(&PersistenceTier::Worker));
         assert!(tiers.contains(&PersistenceTier::EdgeServer));
